Starting rewrite of old digest test cases
